ffi: Create separate timeline object, mirroring the Rust API
